+++ max_div/solver/_presets/preset_random.py
@@ -13,7 +13,7 @@ def get_preset_strategies_random(
     target_duration: TargetDuration,
 ) -> tuple[InitializationStrategy, list[OptimizationStep]]:
     # --- initialization ----------------------------------
-    init_strategy = InitializationStrategy.random_one_shot(uniform=True)
+    init_strategy = InitializationStrategy.fast()
 
     # --- optimization steps ------------------------------
     optim_steps = [