+++ max_div/solver/_presets/preset_smart.py
@@ -1,5 +1,3 @@
-import math
-
 from max_div.solver._duration import TargetDuration
 from max_div.solver._problem import MaxDivProblem
 from max_div.solver._solver_step import OptimizationStep